# docs/swagger.yaml
components:
  schemas:
    files.File:
      properties:
        file:
          format: binary
          type: string
      type: object
...
    models.ErrorResponse:
      properties:
        error:
          example: Message d'erreur
          type: string
        message:
          example: Détail de l'erreur
          type: string
      type: object
...
    models.Response:
      properties:
        data: {}
        message:
          example: Opération réussie
          type: string
        success:
          example: true
          type: boolean
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    name: Équipe Transat
  description: API Backend pour l'application Transat - l'application du campus de
    Nantes d'IMT Atlantique
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: API Transat
  version: "2.0"
openapi: 3.1.0
paths:
...
  /data/{filename}:
    get:
      description: Sert un fichier stocké dans le cloud storage
      parameters:
      - description: Nom du fichier à servir
        in: path
        name: filename
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/files.File'
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Fichier servi avec succès
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Nom de fichier manquant
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Fichier non trouvé
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Erreur serveur
      summary: Servir un fichier
      tags:
      - Files
  /files:
    get:
      description: Récupère la liste de tous les fichiers téléchargés par l'utilisateur
        connecté
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Liste des fichiers récupérée avec succès
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Non autorisé
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Erreur serveur
      security:
      - BearerAuth: []
      summary: Lister les fichiers de l'utilisateur
      tags:
      - Files
  /files/{filename}:
    delete:
      description: Supprime un fichier spécifique appartenant à l'utilisateur connecté
      parameters:
      - description: Nom du fichier à supprimer
        in: path
        name: filename
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Response'
          description: Fichier supprimé avec succès
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Nom de fichier manquant
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Non autorisé
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Fichier non trouvé
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Erreur serveur
      security:
      - BearerAuth: []
      summary: Supprimer un fichier
      tags:
      - Files
...
  /upload:
    post:
      description: Télécharge un fichier et l'enregistre dans le stockage cloud
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/files.File'
        description: Fichier à uploader
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Fichier téléchargé avec succès
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Données invalides
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Non autorisé
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Erreur serveur
      security:
      - BearerAuth: []
      summary: Télécharger un fichier
      tags:
      - Files
...
servers:
- description: Localhost
  url: http://localhost:3000/api
- description: Production
  url: https://api.transat.dev/api
...